Update relabundance.R
The code has been commented on. The function's intent was explained, and a deprecation warning and suggested replacement were also provided. An example of its use was given, and its input and output were described. To further specify which functions are being imported and exported, the @importFrom and @export tags were included.